+++ app/package.json
@@ -24,6 +24,7 @@
     "react-before-after-slider": "^1.0.4",
     "react-bootstrap": "^1.0.0-beta.5",
     "react-bootstrap-daterangepicker": "^4.1.0",
+    "react-coverflow": "^0.2.18",
     "react-dom": "^16.7.0",
     "react-flexbox-grid": "^2.1.2",
     "react-gentelella": "^0.7.3",
