+++ app/src/pages/Diet/QuickSettings.js
