+++ app/src/pages/Diet/RecentActivities.js
