+++ app/src/pages/index.js
@@ -1,10 +1,10 @@
-export { default as AboutPage } from './AboutPage/AboutPage';
+export { default as AboutPage } from './About/AboutPage';
 export { default as ContactPage } from './ContactPage';
-export { default as DietPage } from './DietPage/DietPage';
+export { default as DietPage } from './Diet/DietPage';
 export { default as FitnessPage } from './FitnessPage';
 export { default as HomePage } from './HomePage';
 export { default as LandingPage } from './LandingPage';
 export { default as LivePage } from './LivePage';
 export { default as LocationPage } from './LocationPage';
 export { default as LoginPage } from './LoginPage';
-export { default as MediaPage } from './MediaPage';
+export { default as MediaPage } from './Media/MediaPage';
